CALCULADORA 2.0

+++ TP1/CALCULADORA/Aritmeticas.c
@@ -1,6 +1,10 @@
 #include "Aritmeticas.h"
-
 //FUNCION SUMA
+/** \brief Esta función recibe dos números y los suma.
+ * \param Primer número ingresado para realizar la operación.
+ * \param Segundo número ingresado para realizar la operación.
+ * \return Resultado de la suma.
+ */
 double Sumar(double operando1,double operando2)
 {
     double resultado;
@@ -9,6 +13,11 @@ double Sumar(double operando1,double operando2)
 }
 
 //FUNCION RESTA
+/** \brief Esta función recibe dos números y los resta.
+ * \param Primer número ingresado para realizar la operación.
+ * \param Segundo número ingresado para realizar la operación.
+ * \return Resultado de la resta.
+ */
 double Restar(double operando1,double operando2)
 {
     double resultado;
@@ -17,6 +26,11 @@ double Restar(double operando1,double operando2)
 }
 
 //FUNCION DIVIDIR
+/** \brief Esta función recibe dos números y los divide.
+ * \param Primer número ingresado para realizar la operación.
+ * \param Segundo número ingresado para realizar la operación.
+ * \return Resultado de la división.
+ */
 double Dividir(double operando1,double operando2)
 {
     double resultado;
@@ -25,6 +39,11 @@ double Dividir(double operando1,double operando2)
 }
 
 //FUNCION MULTIPLICAR
+/** \brief Esta función recibe dos números y los multiplica.
+ * \param Primer número ingresado para realizar la operación.
+ * \param Segundo número ingresado para realizar la operación.
+ * \return Resultado de la multiplicación.
+ */
 double Multiplicar(double operando1,double operando2)
 {
     double resultado;
@@ -33,7 +52,11 @@ double Multiplicar(double operando1,double operando2)
 }
 
 //FUNCION FACTORIAL
-double Factorial (double numero)
+/** \brief Esta función realiza el factorial del parametro
+ * \param Número a factorear.
+ * \return Resultado del factorial.
+ */
+double Factorial (int numero)
 {
     double resultado;
     if (numero == 0)
