+++ TP1/CALCULADORA/Aritmeticas.h
@@ -1,9 +1,14 @@
+/*Suma los parametros*/
 double Sumar(double operando1,double operando2);
 
+/*Resta los parametros*/
 double Restar(double operando1,double operando2);
 
+/*Divide los parametros*/
 double Dividir(double operando1,double operando2);
 
+/*Multiplica los parametros*/
 double Multiplicar(double operando1,double operando2);
 
-double Factorial(double numero);
+/*Obtiene el factorial del parámetro*/
+double Factorial(int numero);
